requirements.txt
streamlit>=1.28.0
google-genai>=0.8.0
pandas>=1.5.0
Pillow>=9.0.0 